Refactor: Extract footpath drag slope algorithm to world layer
Moves CalculateConnectedPathSlopes from Footpath window to a separate file in the world layer for better organization
