+++ src/openrct2/libopenrct2.vcxproj
@@ -1276,6 +1276,7 @@
     <ClCompile Include="world\ConstructionClearance.cpp" />
     <ClCompile Include="world\Entrance.cpp" />
     <ClCompile Include="world\Footpath.cpp" />
+    <ClCompile Include="world\FootpathDragSlope.cpp" />
     <ClCompile Include="world\Map.cpp" />
     <ClCompile Include="world\MapAnimation.cpp" />
     <ClCompile Include="world\MapOwnership.cpp" />
